Antigravity-AgenticAIs copy/Agent-2/requirements.txt
langchain-google-genai
langchain
langgraph
python-dotenv
